requirements.txt
anyio==3.6.2
certifi==2022.12.7
charset-normalizer==2.1.1
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
pyaes==1.6.1
pyasn1==0.4.8
python-telegram-bot==20.0
requests==2.28.1
rfc3986==1.5.0
rsa==4.9
sniffio==1.3.0
urllib3==1.26.13
